backend/requirements-minimal.txt
# Minimal requirements for testing and development (without heavy ML dependencies)
# Core web framework
uvicorn[standard]==0.34.2
fastapi==0.115.12
python-multipart==0.0.9
aiofiles

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg

# Basic AI/LLM (without heavy ML packages)
openai
langchain>=0.1.0
langchain-openai>=0.0.5

# Basic workflow
langgraph

# Testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-mock>=3.11.0
httpx>=0.24.0
coverage>=7.2.0

# Utilities
requests
beautifulsoup4
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0

# Development
ipykernel==6.29.5

# NVIDIA Integration
httpx
python-dotenv>=1.0.0
